fix double diffLevel

+++ backend-app/src/main/java/utc/englishlearning/Encybara/domain/Course.java
@@ -32,8 +32,8 @@ public class Course {
     private long id;
     private String name;
     private String intro;
-    private int diffLevel;
-    private int recomLevel;
+    private double diffLevel;
+    private double recomLevel;
     private CourseTypeEnum courseType;
     private SpecialFieldEnum speciField;
     private int numLike;
